requirements.txt
frida
frida-tools
pyaxmlparser
asn1crypto
requests
flask
sqlalchemy
flask-sqlalchemy
python-magic
pwntools
pyOpenSSL
kafka-python
pyjwt
validators
psutil
jsbeautifier
lz4
mitmproxy
psycopg2-binary